Add files via upload
Updated the "dead" synapse controller to use an online estimate of a lognormal quantile using a fixed estimate for the erfcinv function. Uses CV of the weights as an estimate for the lognormal quantiles.
Updated mnist_power file to test a simple working regime (these parameter settings result in feedforward receptive fields with stable long term statistics, and slight drift). The toy model works with this and shows various currently non-trivial behaviors. As said, proof of existence. Code is still very WiP and lacking proper documentation. (more to follow, but the code has some comments and is structured enough to show the network "rules")